+++ integrations/mistral/src/haystack_integrations/components/agents/mistral/__init__.py
@@ -1,4 +1,3 @@
 from .agent import MistralAgent
-from .agent_sdk import MistralAgentSDK
 
-__all__ = ["MistralAgent", "MistralAgentSDK"]
+__all__ = ["MistralAgent"]
