--- integrations/mistral/src/haystack_integrations/components/agents/mistral/agent_sdk.py
@@ -1,640 +1,0 @@
-# SPDX-FileCopyrightText: 2023-present deepset GmbH <info@deepset.ai>
-#
-# SPDX-License-Identifier: Apache-2.0
-
-"""
-MistralAgent using the official Mistral Python SDK.
-
-This implementation uses the `mistralai` package instead of direct HTTP requests,
-providing better error handling, automatic retries, and maintained compatibility
-with the Mistral API.
-
-Requires: pip install mistralai
-"""
-
-import json
-from typing import Any, Literal, Optional, Union
-
-from haystack import component, default_from_dict, default_to_dict, logging
-from haystack.dataclasses import (
-    ChatMessage,
-    StreamingCallbackT,
-    StreamingChunk,
-    ToolCall,
-)
-from haystack.tools import (
-    ToolsType,
-    _check_duplicate_tool_names,
-    deserialize_tools_or_toolset_inplace,
-    flatten_tools_or_toolsets,
-    serialize_tools_or_toolset,
-)
-from haystack.utils import Secret, deserialize_callable, deserialize_secrets_inplace, serialize_callable
-
-logger = logging.getLogger(__name__)
-
-# Type alias for tool_choice parameter
-ToolChoiceType = Union[
-    Literal["auto", "none", "any", "required"],
-    dict[str, Any],
-]
-
-
-@component
-class MistralAgentSDK:
-    """
-    Generates text using Mistral AI Agents via the official Mistral Python SDK.
-
-    This component uses the `mistralai` package to interact with pre-configured
-    Mistral Agents. It provides better error handling and automatic retries
-    compared to direct HTTP requests.
-
-    For more information on Mistral Agents, see:
-    [Mistral Agents API](https://docs.mistral.ai/api/endpoint/agents)
-
-    Prerequisites:
-        pip install mistralai
-
-    Usage example:
-    ```python
-    from haystack_integrations.components.agents.mistral import MistralAgentSDK
-    from haystack.dataclasses import ChatMessage
-
-    # Initialize with your agent ID from the Mistral console
-    agent = MistralAgentSDK(agent_id="your-agent-id")
-
-    messages = [ChatMessage.from_user("What can you help me with?")]
-    response = agent.run(messages)
-    print(response["replies"][0].text)
-    ```
-
-    Streaming example:
-    ```python
-    def my_callback(chunk):
-        print(chunk.content, end="", flush=True)
-
-    agent = MistralAgentSDK(
-        agent_id="your-agent-id",
-        streaming_callback=my_callback
-    )
-    response = agent.run([ChatMessage.from_user("Tell me a story")])
-    ```
-    """
-
-    def __init__(
-        self,
-        agent_id: str,
-        api_key: Secret = Secret.from_env_var("MISTRAL_API_KEY"),
-        streaming_callback: Optional[StreamingCallbackT] = None,
-        tools: Optional[ToolsType] = None,
-        tool_choice: Optional[ToolChoiceType] = None,
-        parallel_tool_calls: bool = True,
-        generation_kwargs: Optional[dict[str, Any]] = None,
-        *,
-        timeout_ms: Optional[int] = 30000,
-        max_retries: int = 3,
-    ):
-        """
-        Creates an instance of MistralAgentSDK.
-
-        :param agent_id:
-            The ID of the Mistral Agent to use. Required. Get this from the
-            Mistral AI console after creating an agent.
-        :param api_key:
-            The Mistral API key. Defaults to environment variable `MISTRAL_API_KEY`.
-        :param streaming_callback:
-            A callback function called when a new token is received from the stream.
-        :param tools:
-            Additional tools the agent can use beyond its pre-configured tools.
-            A list of Tool and/or Toolset objects.
-        :param tool_choice:
-            Controls which tool is called. Options:
-            - "auto": Model decides whether to use tools
-            - "none": No tools, generate text only
-            - "any" or "required": Must call one or more tools
-            - {"type": "function", "function": {"name": "..."}}: Force specific tool
-        :param parallel_tool_calls:
-            Whether to enable parallel function calling. Defaults to True.
-        :param generation_kwargs:
-            Additional parameters for the API call. Supported parameters:
-            - `max_tokens`: Maximum tokens to generate
-            - `frequency_penalty`: Penalize word repetition (default: 0)
-            - `presence_penalty`: Encourage vocabulary diversity (default: 0)
-            - `n`: Number of completions to return
-            - `random_seed`: Seed for deterministic results
-            - `stop`: Stop sequences
-            - `response_format`: Output format (text/json_object/json_schema)
-            - `prediction`: Expected completion for optimization
-            - `prompt_mode`: Set to "reasoning" for reasoning models
-        :param timeout_ms:
-            Request timeout in milliseconds. Defaults to 30000 (30 seconds).
-        :param max_retries:
-            Maximum number of retries on failure. Defaults to 3.
-        """
-        self.agent_id = agent_id
-        self.api_key = api_key
-        self.streaming_callback = streaming_callback
-        self.tools = tools
-        self.tool_choice = tool_choice
-        self.parallel_tool_calls = parallel_tool_calls
-        self.generation_kwargs = generation_kwargs or {}
-        self.timeout_ms = timeout_ms
-        self.max_retries = max_retries
-
-        # Validate tools for duplicates
-        _check_duplicate_tool_names(flatten_tools_or_toolsets(self.tools))
-
-        # Client will be initialized in warm_up()
-        self._client = None
-        self._async_client = None
-
-    def warm_up(self):
-        """
-        Initialize the Mistral SDK client.
-
-        This is called automatically before the first request.
-        """
-        if self._client is not None:
-            return
-
-        try:
-            from mistralai import Mistral
-        except ImportError as e:
-            raise ImportError(
-                "The 'mistralai' package is required for MistralAgentSDK. "
-                "Install it with: pip install mistralai"
-            ) from e
-
-        self._client = Mistral(
-            api_key=self.api_key.resolve_value(),
-            timeout_ms=self.timeout_ms,
-            max_retries=self.max_retries,
-        )
-
-    def _convert_messages_to_sdk_format(
-        self, messages: list[ChatMessage]
-    ) -> list[dict[str, Any]]:
-        """
-        Convert Haystack ChatMessages to Mistral SDK message format.
-
-        The Mistral SDK expects messages as dicts with 'role' and 'content' keys.
-        """
-        sdk_messages = []
-
-        for msg in messages:
-            # Use the OpenAI format which is compatible with Mistral
-            openai_format = msg.to_openai_dict_format()
-
-            # Ensure content is a string (not a list of content blocks)
-            content = openai_format.get("content", "")
-            if isinstance(content, list):
-                # Extract text from content blocks
-                text_parts = []
-                for block in content:
-                    if isinstance(block, dict) and block.get("type") == "text":
-                        text_parts.append(block.get("text", ""))
-                    elif isinstance(block, str):
-                        text_parts.append(block)
-                content = "".join(text_parts)
-
-            sdk_message = {
-                "role": openai_format.get("role", "user"),
-                "content": content,
-            }
-
-            # Include tool_call_id for tool messages
-            if openai_format.get("tool_call_id"):
-                sdk_message["tool_call_id"] = openai_format["tool_call_id"]
-
-            # Include tool_calls for assistant messages
-            if openai_format.get("tool_calls"):
-                sdk_message["tool_calls"] = openai_format["tool_calls"]
-
-            sdk_messages.append(sdk_message)
-
-        return sdk_messages
-
-    def _build_tools_for_sdk(
-        self, tools: Optional[ToolsType] = None
-    ) -> Optional[list[dict[str, Any]]]:
-        """Convert Haystack tools to Mistral SDK format."""
-        flattened_tools = flatten_tools_or_toolsets(tools or self.tools)
-        if not flattened_tools:
-            return None
-
-        return [
-            {"type": "function", "function": tool.tool_spec}
-            for tool in flattened_tools
-        ]
-
-    def _parse_sdk_response(self, response: Any) -> list[ChatMessage]:
-        """
-        Parse the Mistral SDK response into Haystack ChatMessages.
-
-        :param response: The response from mistral.agents.complete()
-        :returns: List of ChatMessage objects
-        """
-        messages = []
-
-        for choice in response.choices:
-            message = choice.message
-            content = message.content or ""
-
-            # Parse tool calls if present
-            tool_calls = []
-            if message.tool_calls:
-                for tc in message.tool_calls:
-                    if tc.type == "function":
-                        try:
-                            arguments = json.loads(tc.function.arguments or "{}")
-                        except json.JSONDecodeError:
-                            arguments = {}
-                        tool_calls.append(
-                            ToolCall(
-                                id=tc.id,
-                                tool_name=tc.function.name,
-                                arguments=arguments,
-                            )
-                        )
-
-            # Build metadata
-            meta = {
-                "model": response.model,
-                "index": choice.index,
-                "finish_reason": choice.finish_reason,
-                "usage": {
-                    "prompt_tokens": response.usage.prompt_tokens,
-                    "completion_tokens": response.usage.completion_tokens,
-                    "total_tokens": response.usage.total_tokens,
-                } if response.usage else None,
-            }
-
-            chat_message = ChatMessage.from_assistant(
-                text=content if content else None,
-                tool_calls=tool_calls if tool_calls else None,
-                meta=meta,
-            )
-            messages.append(chat_message)
-
-        return messages
-
-    def _handle_streaming(
-        self,
-        stream_response: Any,
-        callback: StreamingCallbackT,
-    ) -> list[ChatMessage]:
-        """
-        Handle streaming response from the Mistral SDK.
-
-        :param stream_response: The streaming response iterator
-        :param callback: The callback to invoke for each chunk
-        :returns: List containing the final assembled ChatMessage
-        """
-        collected_content = ""
-        collected_tool_calls: dict[int, dict] = {}
-        meta: dict[str, Any] = {}
-
-        for chunk in stream_response:
-            # Extract metadata from response
-            if not meta and chunk.model:
-                meta["model"] = chunk.model
-
-            for choice in chunk.data.choices:
-                delta = choice.delta
-
-                # Handle text content
-                if delta.content:
-                    collected_content += delta.content
-                    streaming_chunk = StreamingChunk(
-                        content=delta.content,
-                        meta={
-                            "model": chunk.model,
-                            "index": choice.index,
-                            "finish_reason": choice.finish_reason,
-                        },
-                    )
-                    callback(streaming_chunk)
-
-                # Handle tool calls
-                if delta.tool_calls:
-                    for tc in delta.tool_calls:
-                        idx = tc.index if hasattr(tc, "index") else 0
-                        if idx not in collected_tool_calls:
-                            collected_tool_calls[idx] = {
-                                "id": "",
-                                "name": "",
-                                "arguments": "",
-                            }
-                        if tc.id:
-                            collected_tool_calls[idx]["id"] = tc.id
-                        if hasattr(tc, "function") and tc.function:
-                            if tc.function.name:
-                                collected_tool_calls[idx]["name"] = tc.function.name
-                            if tc.function.arguments:
-                                collected_tool_calls[idx]["arguments"] += tc.function.arguments
-
-                # Capture finish reason
-                if choice.finish_reason:
-                    meta["finish_reason"] = choice.finish_reason
-                    meta["index"] = choice.index
-
-            # Capture usage from final chunk
-            if chunk.data.usage:
-                meta["usage"] = {
-                    "prompt_tokens": chunk.data.usage.prompt_tokens,
-                    "completion_tokens": chunk.data.usage.completion_tokens,
-                    "total_tokens": chunk.data.usage.total_tokens,
-                }
-
-        # Build final tool calls
-        tool_calls = []
-        for idx in sorted(collected_tool_calls.keys()):
-            tc_data = collected_tool_calls[idx]
-            try:
-                arguments = json.loads(tc_data["arguments"]) if tc_data["arguments"] else {}
-            except json.JSONDecodeError:
-                arguments = {}
-            tool_calls.append(
-                ToolCall(
-                    id=tc_data["id"],
-                    tool_name=tc_data["name"],
-                    arguments=arguments,
-                )
-            )
-
-        # Create final message
-        chat_message = ChatMessage.from_assistant(
-            text=collected_content if collected_content else None,
-            tool_calls=tool_calls if tool_calls else None,
-            meta=meta,
-        )
-
-        return [chat_message]
-
-    @component.output_types(replies=list[ChatMessage])
-    def run(
-        self,
-        messages: list[ChatMessage],
-        streaming_callback: Optional[StreamingCallbackT] = None,
-        tools: Optional[ToolsType] = None,
-        tool_choice: Optional[ToolChoiceType] = None,
-        generation_kwargs: Optional[dict[str, Any]] = None,
-    ) -> dict[str, list[ChatMessage]]:
-        """
-        Invoke the Mistral Agent with the provided messages.
-
-        :param messages:
-            A list of ChatMessage instances representing the conversation.
-        :param streaming_callback:
-            A callback function for streaming. Overrides the init callback.
-        :param tools:
-            Additional tools for this request. Overrides init tools.
-        :param tool_choice:
-            Tool choice for this request. Overrides init tool_choice.
-        :param generation_kwargs:
-            Additional generation parameters. Merged with init params.
-
-        :returns:
-            A dictionary with key `replies` containing a list of ChatMessage responses.
-        """
-        self.warm_up()
-
-        if not messages:
-            return {"replies": []}
-
-        # Select streaming callback
-        effective_callback = streaming_callback or self.streaming_callback
-
-        # Convert messages
-        sdk_messages = self._convert_messages_to_sdk_format(messages)
-
-        # Merge generation kwargs
-        merged_kwargs = {**self.generation_kwargs, **(generation_kwargs or {})}
-
-        # Build request kwargs
-        request_kwargs: dict[str, Any] = {
-            "agent_id": self.agent_id,
-            "messages": sdk_messages,
-        }
-
-        # Add tools if provided
-        sdk_tools = self._build_tools_for_sdk(tools)
-        if sdk_tools:
-            request_kwargs["tools"] = sdk_tools
-            request_kwargs["parallel_tool_calls"] = self.parallel_tool_calls
-
-            # Add tool_choice only when tools are present
-            effective_tool_choice = tool_choice or self.tool_choice
-            if effective_tool_choice:
-                request_kwargs["tool_choice"] = effective_tool_choice
-
-        # Add generation kwargs
-        for key, value in merged_kwargs.items():
-            if value is not None:
-                request_kwargs[key] = value
-
-        try:
-            if effective_callback:
-                # Streaming request
-                request_kwargs["stream"] = True
-                stream_response = self._client.agents.stream(**request_kwargs)
-                replies = self._handle_streaming(stream_response, effective_callback)
-            else:
-                # Non-streaming request
-                response = self._client.agents.complete(**request_kwargs)
-                replies = self._parse_sdk_response(response)
-
-            return {"replies": replies}
-
-        except Exception as e:
-            # Import error types from mistralai
-            try:
-                from mistralai import models
-
-                if isinstance(e, models.HTTPValidationError):
-                    logger.error(
-                        "Mistral validation error: {detail}",
-                        detail=e.data.detail if hasattr(e, "data") else str(e),
-                    )
-                    raise ValueError(f"Mistral validation error: {e}") from e
-                elif isinstance(e, models.MistralError):
-                    logger.error(
-                        "Mistral API error: {status_code} - {message}",
-                        status_code=e.status_code,
-                        message=e.message,
-                    )
-                    raise ValueError(f"Mistral API error ({e.status_code}): {e.message}") from e
-            except ImportError:
-                pass
-            raise
-
-    @component.output_types(replies=list[ChatMessage])
-    async def run_async(
-        self,
-        messages: list[ChatMessage],
-        streaming_callback: Optional[StreamingCallbackT] = None,
-        tools: Optional[ToolsType] = None,
-        tool_choice: Optional[ToolChoiceType] = None,
-        generation_kwargs: Optional[dict[str, Any]] = None,
-    ) -> dict[str, list[ChatMessage]]:
-        """
-        Asynchronously invoke the Mistral Agent.
-
-        Same parameters as `run()`.
-        """
-        self.warm_up()
-
-        if not messages:
-            return {"replies": []}
-
-        effective_callback = streaming_callback or self.streaming_callback
-        sdk_messages = self._convert_messages_to_sdk_format(messages)
-        merged_kwargs = {**self.generation_kwargs, **(generation_kwargs or {})}
-
-        request_kwargs: dict[str, Any] = {
-            "agent_id": self.agent_id,
-            "messages": sdk_messages,
-        }
-
-        sdk_tools = self._build_tools_for_sdk(tools)
-        if sdk_tools:
-            request_kwargs["tools"] = sdk_tools
-            request_kwargs["parallel_tool_calls"] = self.parallel_tool_calls
-            effective_tool_choice = tool_choice or self.tool_choice
-            if effective_tool_choice:
-                request_kwargs["tool_choice"] = effective_tool_choice
-
-        for key, value in merged_kwargs.items():
-            if value is not None:
-                request_kwargs[key] = value
-
-        try:
-            if effective_callback:
-                # Async streaming
-                request_kwargs["stream"] = True
-                stream_response = await self._client.agents.stream_async(**request_kwargs)
-                # Note: For full async streaming, we'd need an async callback
-                # This is a simplified version
-                replies = await self._handle_async_streaming(stream_response, effective_callback)
-            else:
-                response = await self._client.agents.complete_async(**request_kwargs)
-                replies = self._parse_sdk_response(response)
-
-            return {"replies": replies}
-
-        except Exception as e:
-            try:
-                from mistralai import models
-                if isinstance(e, (models.HTTPValidationError, models.MistralError)):
-                    raise ValueError(f"Mistral API error: {e}") from e
-            except ImportError:
-                pass
-            raise
-
-    async def _handle_async_streaming(
-        self,
-        stream_response: Any,
-        callback: StreamingCallbackT,
-    ) -> list[ChatMessage]:
-        """Handle async streaming response."""
-        collected_content = ""
-        collected_tool_calls: dict[int, dict] = {}
-        meta: dict[str, Any] = {}
-
-        async for chunk in stream_response:
-            if not meta and chunk.model:
-                meta["model"] = chunk.model
-
-            for choice in chunk.data.choices:
-                delta = choice.delta
-
-                if delta.content:
-                    collected_content += delta.content
-                    streaming_chunk = StreamingChunk(
-                        content=delta.content,
-                        meta={
-                            "model": chunk.model,
-                            "index": choice.index,
-                            "finish_reason": choice.finish_reason,
-                        },
-                    )
-                    # For async streaming, callback should be awaitable
-                    if callable(callback):
-                        result = callback(streaming_chunk)
-                        if hasattr(result, "__await__"):
-                            await result
-
-                if delta.tool_calls:
-                    for tc in delta.tool_calls:
-                        idx = tc.index if hasattr(tc, "index") else 0
-                        if idx not in collected_tool_calls:
-                            collected_tool_calls[idx] = {"id": "", "name": "", "arguments": ""}
-                        if tc.id:
-                            collected_tool_calls[idx]["id"] = tc.id
-                        if hasattr(tc, "function") and tc.function:
-                            if tc.function.name:
-                                collected_tool_calls[idx]["name"] = tc.function.name
-                            if tc.function.arguments:
-                                collected_tool_calls[idx]["arguments"] += tc.function.arguments
-
-                if choice.finish_reason:
-                    meta["finish_reason"] = choice.finish_reason
-                    meta["index"] = choice.index
-
-            if chunk.data.usage:
-                meta["usage"] = {
-                    "prompt_tokens": chunk.data.usage.prompt_tokens,
-                    "completion_tokens": chunk.data.usage.completion_tokens,
-                    "total_tokens": chunk.data.usage.total_tokens,
-                }
-
-        tool_calls = []
-        for idx in sorted(collected_tool_calls.keys()):
-            tc_data = collected_tool_calls[idx]
-            try:
-                arguments = json.loads(tc_data["arguments"]) if tc_data["arguments"] else {}
-            except json.JSONDecodeError:
-                arguments = {}
-            tool_calls.append(
-                ToolCall(id=tc_data["id"], tool_name=tc_data["name"], arguments=arguments)
-            )
-
-        return [
-            ChatMessage.from_assistant(
-                text=collected_content if collected_content else None,
-                tool_calls=tool_calls if tool_calls else None,
-                meta=meta,
-            )
-        ]
-
-    def to_dict(self) -> dict[str, Any]:
-        """Serialize this component to a dictionary."""
-        callback_name = serialize_callable(self.streaming_callback) if self.streaming_callback else None
-
-        return default_to_dict(
-            self,
-            agent_id=self.agent_id,
-            api_key=self.api_key.to_dict(),
-            streaming_callback=callback_name,
-            tools=serialize_tools_or_toolset(self.tools),
-            tool_choice=self.tool_choice,
-            parallel_tool_calls=self.parallel_tool_calls,
-            generation_kwargs=self.generation_kwargs,
-            timeout_ms=self.timeout_ms,
-            max_retries=self.max_retries,
-        )
-
-    @classmethod
-    def from_dict(cls, data: dict[str, Any]) -> "MistralAgentSDK":
-        """Deserialize this component from a dictionary."""
-        deserialize_secrets_inplace(data["init_parameters"], keys=["api_key"])
-        deserialize_tools_or_toolset_inplace(data["init_parameters"], key="tools")
-
-        init_params = data.get("init_parameters", {})
-        if init_params.get("streaming_callback"):
-            data["init_parameters"]["streaming_callback"] = deserialize_callable(
-                init_params["streaming_callback"]
-            )
-
-        return default_from_dict(cls, data)
-
