--- integrations/mistral/tests/test_mistral_agent_sdk.py
@@ -1,733 +1,0 @@
-# SPDX-FileCopyrightText: 2023-present deepset GmbH <info@deepset.ai>
-#
-# SPDX-License-Identifier: Apache-2.0
-
-"""
-Unit tests for MistralAgentSDK component.
-
-These tests mock the Mistral SDK client to test the component logic
-without making actual API calls.
-"""
-
-import json
-from unittest.mock import MagicMock, patch, AsyncMock
-from typing import Any
-
-import pytest
-from haystack import Pipeline
-from haystack.dataclasses import ChatMessage, ChatRole, StreamingChunk, ToolCall
-from haystack.tools import Tool, Toolset
-from haystack.utils.auth import Secret
-
-from haystack_integrations.components.agents.mistral.agent_sdk import MistralAgentSDK
-
-
-# =============================================================================
-# FIXTURES
-# =============================================================================
-
-
-class MockUsage:
-    """Mock usage object from Mistral SDK."""
-    def __init__(self, prompt_tokens: int = 10, completion_tokens: int = 20, total_tokens: int = 30):
-        self.prompt_tokens = prompt_tokens
-        self.completion_tokens = completion_tokens
-        self.total_tokens = total_tokens
-
-
-class MockFunction:
-    """Mock function object for tool calls."""
-    def __init__(self, name: str, arguments: str):
-        self.name = name
-        self.arguments = arguments
-
-
-class MockToolCall:
-    """Mock tool call object from Mistral SDK."""
-    def __init__(self, id: str, name: str, arguments: str, type: str = "function"):
-        self.id = id
-        self.type = type
-        self.function = MockFunction(name=name, arguments=arguments)
-
-
-class MockMessage:
-    """Mock message object from Mistral SDK response."""
-    def __init__(self, content: str = "", tool_calls: list = None):
-        self.content = content
-        self.tool_calls = tool_calls or []
-
-
-class MockChoice:
-    """Mock choice object from Mistral SDK response."""
-    def __init__(self, message: MockMessage, index: int = 0, finish_reason: str = "stop"):
-        self.message = message
-        self.index = index
-        self.finish_reason = finish_reason
-
-
-class MockResponse:
-    """Mock response object from Mistral SDK."""
-    def __init__(self, choices: list, model: str = "agent-model", usage: MockUsage = None):
-        self.choices = choices
-        self.model = model
-        self.usage = usage or MockUsage()
-
-
-class MockDelta:
-    """Mock delta object for streaming."""
-    def __init__(self, content: str = None, tool_calls: list = None):
-        self.content = content
-        self.tool_calls = tool_calls
-
-
-class MockStreamChoice:
-    """Mock streaming choice object."""
-    def __init__(self, delta: MockDelta, index: int = 0, finish_reason: str = None):
-        self.delta = delta
-        self.index = index
-        self.finish_reason = finish_reason
-
-
-class MockStreamData:
-    """Mock streaming data object."""
-    def __init__(self, choices: list, usage: MockUsage = None):
-        self.choices = choices
-        self.usage = usage
-
-
-class MockStreamChunk:
-    """Mock streaming chunk from SDK."""
-    def __init__(self, data: MockStreamData, model: str = "agent-model"):
-        self.data = data
-        self.model = model
-
-
-@pytest.fixture
-def chat_messages():
-    """Basic chat messages for testing."""
-    return [
-        ChatMessage.from_system("You are a helpful assistant"),
-        ChatMessage.from_user("What's the capital of France?"),
-    ]
-
-
-@pytest.fixture
-def mock_tool_function():
-    """A mock tool function."""
-    def weather(city: str) -> str:
-        return f"Weather in {city}: Sunny, 22°C"
-    return weather
-
-
-@pytest.fixture
-def tools(mock_tool_function):
-    """Sample tools for testing."""
-    return [
-        Tool(
-            name="weather",
-            description="Get weather for a city",
-            parameters={
-                "type": "object",
-                "properties": {"city": {"type": "string"}},
-                "required": ["city"]
-            },
-            function=mock_tool_function,
-        )
-    ]
-
-
-@pytest.fixture
-def mock_sdk_response():
-    """Create a mock SDK response."""
-    return MockResponse(
-        choices=[
-            MockChoice(
-                message=MockMessage(content="The capital of France is Paris."),
-                finish_reason="stop",
-                index=0,
-            )
-        ],
-        model="mistral-agent-model",
-        usage=MockUsage(prompt_tokens=15, completion_tokens=25, total_tokens=40),
-    )
-
-
-@pytest.fixture
-def mock_sdk_response_with_tool_call():
-    """Create a mock SDK response with tool calls."""
-    tool_call = MockToolCall(
-        id="call_123",
-        name="weather",
-        arguments='{"city": "Paris"}',
-    )
-    return MockResponse(
-        choices=[
-            MockChoice(
-                message=MockMessage(content="", tool_calls=[tool_call]),
-                finish_reason="tool_calls",
-                index=0,
-            )
-        ],
-        model="mistral-agent-model",
-        usage=MockUsage(prompt_tokens=20, completion_tokens=15, total_tokens=35),
-    )
-
-
-# =============================================================================
-# INITIALIZATION TESTS
-# =============================================================================
-
-
-class TestMistralAgentSDKInit:
-    """Tests for MistralAgentSDK initialization."""
-
-    def test_init_default(self, monkeypatch):
-        """Test default initialization."""
-        monkeypatch.setenv("MISTRAL_API_KEY", "test-api-key")
-
-        agent = MistralAgentSDK(agent_id="ag-test-123")
-
-        assert agent.agent_id == "ag-test-123"
-        assert agent.streaming_callback is None
-        assert agent.tools is None
-        assert agent.tool_choice is None
-        assert agent.parallel_tool_calls is True
-        assert agent.generation_kwargs == {}
-        assert agent.timeout_ms == 30000
-        assert agent.max_retries == 3
-
-    def test_init_with_parameters(self, tools):
-        """Test initialization with custom parameters."""
-        def my_callback(chunk):
-            pass
-
-        agent = MistralAgentSDK(
-            agent_id="ag-custom-456",
-            api_key=Secret.from_token("custom-key"),
-            streaming_callback=my_callback,
-            tools=tools,
-            tool_choice="auto",
-            parallel_tool_calls=False,
-            generation_kwargs={"max_tokens": 500, "random_seed": 42},
-            timeout_ms=60000,
-            max_retries=5,
-        )
-
-        assert agent.agent_id == "ag-custom-456"
-        assert agent.streaming_callback is my_callback
-        assert len(agent.tools) == 1
-        assert agent.tool_choice == "auto"
-        assert agent.parallel_tool_calls is False
-        assert agent.generation_kwargs == {"max_tokens": 500, "random_seed": 42}
-        assert agent.timeout_ms == 60000
-        assert agent.max_retries == 5
-
-    def test_init_fail_without_api_key(self, monkeypatch):
-        """Test that initialization fails without API key."""
-        monkeypatch.delenv("MISTRAL_API_KEY", raising=False)
-
-        with pytest.raises(ValueError, match=r"None of the .* environment variables are set"):
-            MistralAgentSDK(agent_id="ag-test-123")
-
-    def test_init_with_duplicate_tools_raises_error(self, mock_tool_function):
-        """Test that duplicate tool names raise an error."""
-        duplicate_tools = [
-            Tool(name="weather", description="First", parameters={}, function=mock_tool_function),
-            Tool(name="weather", description="Second", parameters={}, function=mock_tool_function),
-        ]
-
-        with pytest.raises(ValueError, match="Duplicate tool names"):
-            MistralAgentSDK(
-                agent_id="ag-test",
-                api_key=Secret.from_token("key"),
-                tools=duplicate_tools,
-            )
-
-    def test_init_with_toolset(self, mock_tool_function, monkeypatch):
-        """Test initialization with a Toolset."""
-        monkeypatch.setenv("MISTRAL_API_KEY", "test-key")
-
-        tool1 = Tool(name="tool1", description="First tool", parameters={}, function=mock_tool_function)
-        tool2 = Tool(name="tool2", description="Second tool", parameters={}, function=mock_tool_function)
-        toolset = Toolset([tool1, tool2])
-
-        agent = MistralAgentSDK(agent_id="ag-test", tools=[toolset])
-
-        assert agent.tools == [toolset]
-
-
-# =============================================================================
-# MESSAGE CONVERSION TESTS
-# =============================================================================
-
-
-class TestMessageConversion:
-    """Tests for message format conversion."""
-
-    def test_convert_simple_user_message(self, monkeypatch):
-        """Test converting a simple user message."""
-        monkeypatch.setenv("MISTRAL_API_KEY", "test-key")
-        agent = MistralAgentSDK(agent_id="ag-test")
-
-        messages = [ChatMessage.from_user("Hello")]
-        result = agent._convert_messages_to_sdk_format(messages)
-
-        assert len(result) == 1
-        assert result[0]["role"] == "user"
-        assert result[0]["content"] == "Hello"
-
-    def test_convert_system_message(self, monkeypatch):
-        """Test converting a system message."""
-        monkeypatch.setenv("MISTRAL_API_KEY", "test-key")
-        agent = MistralAgentSDK(agent_id="ag-test")
-
-        messages = [ChatMessage.from_system("You are helpful")]
-        result = agent._convert_messages_to_sdk_format(messages)
-
-        assert len(result) == 1
-        assert result[0]["role"] == "system"
-        assert result[0]["content"] == "You are helpful"
-
-    def test_convert_assistant_message(self, monkeypatch):
-        """Test converting an assistant message."""
-        monkeypatch.setenv("MISTRAL_API_KEY", "test-key")
-        agent = MistralAgentSDK(agent_id="ag-test")
-
-        messages = [ChatMessage.from_assistant("I can help you")]
-        result = agent._convert_messages_to_sdk_format(messages)
-
-        assert len(result) == 1
-        assert result[0]["role"] == "assistant"
-        assert result[0]["content"] == "I can help you"
-
-    def test_convert_tool_message(self, monkeypatch):
-        """Test converting a tool result message."""
-        monkeypatch.setenv("MISTRAL_API_KEY", "test-key")
-        agent = MistralAgentSDK(agent_id="ag-test")
-
-        tool_call = ToolCall(id="call_123", tool_name="weather", arguments={"city": "Paris"})
-        messages = [ChatMessage.from_tool(tool_result="Sunny, 22°C", origin=tool_call)]
-        result = agent._convert_messages_to_sdk_format(messages)
-
-        assert len(result) == 1
-        assert result[0]["role"] == "tool"
-        assert result[0]["content"] == "Sunny, 22°C"
-        assert result[0]["tool_call_id"] == "call_123"
-
-
-# =============================================================================
-# RUN TESTS (with mocked SDK)
-# =============================================================================
-
-
-class TestMistralAgentSDKRun:
-    """Tests for the run method with mocked SDK."""
-
-    @patch("haystack_integrations.components.agents.mistral.agent_sdk.MistralAgentSDK.warm_up")
-    def test_run_basic(self, mock_warm_up, chat_messages, mock_sdk_response, monkeypatch):
-        """Test basic run without streaming."""
-        monkeypatch.setenv("MISTRAL_API_KEY", "test-key")
-
-        agent = MistralAgentSDK(agent_id="ag-test-123")
-        agent._client = MagicMock()
-        agent._client.agents.complete.return_value = mock_sdk_response
-
-        result = agent.run(chat_messages)
-
-        assert "replies" in result
-        assert len(result["replies"]) == 1
-
-        reply = result["replies"][0]
-        assert reply.text == "The capital of France is Paris."
-        assert reply.meta["model"] == "mistral-agent-model"
-        assert reply.meta["finish_reason"] == "stop"
-        assert reply.meta["usage"]["total_tokens"] == 40
-
-    @patch("haystack_integrations.components.agents.mistral.agent_sdk.MistralAgentSDK.warm_up")
-    def test_run_empty_messages(self, mock_warm_up, monkeypatch):
-        """Test run with empty messages returns empty replies."""
-        monkeypatch.setenv("MISTRAL_API_KEY", "test-key")
-
-        agent = MistralAgentSDK(agent_id="ag-test")
-        result = agent.run([])
-
-        assert result == {"replies": []}
-
-    @patch("haystack_integrations.components.agents.mistral.agent_sdk.MistralAgentSDK.warm_up")
-    def test_run_with_tool_calls(
-        self, mock_warm_up, chat_messages, mock_sdk_response_with_tool_call, tools, monkeypatch
-    ):
-        """Test run that results in tool calls."""
-        monkeypatch.setenv("MISTRAL_API_KEY", "test-key")
-
-        agent = MistralAgentSDK(agent_id="ag-test", tools=tools)
-        agent._client = MagicMock()
-        agent._client.agents.complete.return_value = mock_sdk_response_with_tool_call
-
-        result = agent.run([ChatMessage.from_user("What's the weather in Paris?")])
-
-        assert len(result["replies"]) == 1
-        reply = result["replies"][0]
-
-        assert reply.tool_calls is not None
-        assert len(reply.tool_calls) == 1
-        assert reply.tool_calls[0].tool_name == "weather"
-        assert reply.tool_calls[0].arguments == {"city": "Paris"}
-        assert reply.tool_calls[0].id == "call_123"
-        assert reply.meta["finish_reason"] == "tool_calls"
-
-    @patch("haystack_integrations.components.agents.mistral.agent_sdk.MistralAgentSDK.warm_up")
-    def test_run_with_generation_kwargs(self, mock_warm_up, chat_messages, mock_sdk_response, monkeypatch):
-        """Test run with generation kwargs."""
-        monkeypatch.setenv("MISTRAL_API_KEY", "test-key")
-
-        agent = MistralAgentSDK(
-            agent_id="ag-test",
-            generation_kwargs={"max_tokens": 100}
-        )
-        agent._client = MagicMock()
-        agent._client.agents.complete.return_value = mock_sdk_response
-
-        # Run with additional kwargs
-        agent.run(chat_messages, generation_kwargs={"random_seed": 42})
-
-        # Verify the call was made with merged kwargs
-        call_kwargs = agent._client.agents.complete.call_args.kwargs
-        assert call_kwargs["max_tokens"] == 100
-        assert call_kwargs["random_seed"] == 42
-
-    @patch("haystack_integrations.components.agents.mistral.agent_sdk.MistralAgentSDK.warm_up")
-    def test_run_with_tool_choice(self, mock_warm_up, chat_messages, mock_sdk_response, tools, monkeypatch):
-        """Test run with tool_choice parameter."""
-        monkeypatch.setenv("MISTRAL_API_KEY", "test-key")
-
-        agent = MistralAgentSDK(agent_id="ag-test", tools=tools, tool_choice="any")
-        agent._client = MagicMock()
-        agent._client.agents.complete.return_value = mock_sdk_response
-
-        agent.run(chat_messages)
-
-        call_kwargs = agent._client.agents.complete.call_args.kwargs
-        assert call_kwargs["tool_choice"] == "any"
-        assert "tools" in call_kwargs
-        assert call_kwargs["parallel_tool_calls"] is True
-
-
-# =============================================================================
-# STREAMING TESTS
-# =============================================================================
-
-
-class TestMistralAgentSDKStreaming:
-    """Tests for streaming functionality."""
-
-    @patch("haystack_integrations.components.agents.mistral.agent_sdk.MistralAgentSDK.warm_up")
-    def test_run_with_streaming(self, mock_warm_up, chat_messages, monkeypatch):
-        """Test run with streaming callback."""
-        monkeypatch.setenv("MISTRAL_API_KEY", "test-key")
-
-        collected_chunks = []
-
-        def callback(chunk: StreamingChunk):
-            collected_chunks.append(chunk)
-
-        # Create mock streaming response
-        stream_chunks = [
-            MockStreamChunk(
-                data=MockStreamData(
-                    choices=[MockStreamChoice(delta=MockDelta(content="The "))]
-                ),
-                model="agent-model"
-            ),
-            MockStreamChunk(
-                data=MockStreamData(
-                    choices=[MockStreamChoice(delta=MockDelta(content="capital "))]
-                ),
-                model="agent-model"
-            ),
-            MockStreamChunk(
-                data=MockStreamData(
-                    choices=[MockStreamChoice(delta=MockDelta(content="is Paris."), finish_reason="stop")],
-                    usage=MockUsage(prompt_tokens=10, completion_tokens=5, total_tokens=15)
-                ),
-                model="agent-model"
-            ),
-        ]
-
-        agent = MistralAgentSDK(agent_id="ag-test", streaming_callback=callback)
-        agent._client = MagicMock()
-        agent._client.agents.stream.return_value = iter(stream_chunks)
-
-        result = agent.run(chat_messages)
-
-        # Verify streaming callback was called
-        assert len(collected_chunks) == 3
-        assert collected_chunks[0].content == "The "
-        assert collected_chunks[1].content == "capital "
-        assert collected_chunks[2].content == "is Paris."
-
-        # Verify final message
-        assert len(result["replies"]) == 1
-        assert result["replies"][0].text == "The capital is Paris."
-        assert result["replies"][0].meta["finish_reason"] == "stop"
-
-    @patch("haystack_integrations.components.agents.mistral.agent_sdk.MistralAgentSDK.warm_up")
-    def test_streaming_with_tool_calls(self, mock_warm_up, monkeypatch):
-        """Test streaming response with tool calls."""
-        monkeypatch.setenv("MISTRAL_API_KEY", "test-key")
-
-        collected_chunks = []
-
-        def callback(chunk: StreamingChunk):
-            collected_chunks.append(chunk)
-
-        # Mock tool call delta
-        class MockToolCallDelta:
-            def __init__(self, index, id=None, function=None):
-                self.index = index
-                self.id = id
-                self.function = function
-
-        stream_chunks = [
-            MockStreamChunk(
-                data=MockStreamData(
-                    choices=[
-                        MockStreamChoice(
-                            delta=MockDelta(
-                                content=None,
-                                tool_calls=[
-                                    MockToolCallDelta(
-                                        index=0,
-                                        id="call_abc",
-                                        function=MockFunction(name="weather", arguments='{"city"')
-                                    )
-                                ]
-                            )
-                        )
-                    ]
-                ),
-                model="agent-model"
-            ),
-            MockStreamChunk(
-                data=MockStreamData(
-                    choices=[
-                        MockStreamChoice(
-                            delta=MockDelta(
-                                content=None,
-                                tool_calls=[
-                                    MockToolCallDelta(
-                                        index=0,
-                                        id=None,
-                                        function=MockFunction(name=None, arguments=': "Paris"}')
-                                    )
-                                ]
-                            ),
-                            finish_reason="tool_calls"
-                        )
-                    ],
-                    usage=MockUsage()
-                ),
-                model="agent-model"
-            ),
-        ]
-
-        agent = MistralAgentSDK(agent_id="ag-test", streaming_callback=callback)
-        agent._client = MagicMock()
-        agent._client.agents.stream.return_value = iter(stream_chunks)
-
-        result = agent.run([ChatMessage.from_user("Weather in Paris?")])
-
-        reply = result["replies"][0]
-        assert reply.tool_calls is not None
-        assert len(reply.tool_calls) == 1
-        assert reply.tool_calls[0].tool_name == "weather"
-        assert reply.tool_calls[0].arguments == {"city": "Paris"}
-
-
-# =============================================================================
-# SERIALIZATION TESTS
-# =============================================================================
-
-
-class TestMistralAgentSDKSerialization:
-    """Tests for serialization and deserialization."""
-
-    def test_to_dict_default(self, monkeypatch):
-        """Test serialization with default parameters."""
-        monkeypatch.setenv("MISTRAL_API_KEY", "test-key")
-
-        agent = MistralAgentSDK(agent_id="ag-test-123")
-        data = agent.to_dict()
-
-        assert data["type"] == "haystack_integrations.components.agents.mistral.agent_sdk.MistralAgentSDK"
-        assert data["init_parameters"]["agent_id"] == "ag-test-123"
-        assert data["init_parameters"]["api_key"] == {
-            "env_vars": ["MISTRAL_API_KEY"],
-            "strict": True,
-            "type": "env_var"
-        }
-        assert data["init_parameters"]["streaming_callback"] is None
-        assert data["init_parameters"]["tools"] is None
-        assert data["init_parameters"]["tool_choice"] is None
-        assert data["init_parameters"]["parallel_tool_calls"] is True
-        assert data["init_parameters"]["generation_kwargs"] == {}
-        assert data["init_parameters"]["timeout_ms"] == 30000
-        assert data["init_parameters"]["max_retries"] == 3
-
-    def test_to_dict_with_tools(self, tools, monkeypatch):
-        """Test serialization with tools."""
-        monkeypatch.setenv("MISTRAL_API_KEY", "test-key")
-
-        agent = MistralAgentSDK(
-            agent_id="ag-test",
-            tools=tools,
-            tool_choice="auto",
-            generation_kwargs={"max_tokens": 500}
-        )
-        data = agent.to_dict()
-
-        assert data["init_parameters"]["tool_choice"] == "auto"
-        assert data["init_parameters"]["generation_kwargs"] == {"max_tokens": 500}
-        assert len(data["init_parameters"]["tools"]) == 1
-        assert data["init_parameters"]["tools"][0]["type"] == "haystack.tools.tool.Tool"
-
-    def test_from_dict(self, monkeypatch):
-        """Test deserialization."""
-        monkeypatch.setenv("MISTRAL_API_KEY", "test-key")
-
-        data = {
-            "type": "haystack_integrations.components.agents.mistral.agent_sdk.MistralAgentSDK",
-            "init_parameters": {
-                "agent_id": "ag-restored",
-                "api_key": {"env_vars": ["MISTRAL_API_KEY"], "strict": True, "type": "env_var"},
-                "streaming_callback": None,
-                "tools": None,
-                "tool_choice": "auto",
-                "parallel_tool_calls": False,
-                "generation_kwargs": {"max_tokens": 200},
-                "timeout_ms": 45000,
-                "max_retries": 5,
-            }
-        }
-
-        agent = MistralAgentSDK.from_dict(data)
-
-        assert agent.agent_id == "ag-restored"
-        assert agent.tool_choice == "auto"
-        assert agent.parallel_tool_calls is False
-        assert agent.generation_kwargs == {"max_tokens": 200}
-        assert agent.timeout_ms == 45000
-        assert agent.max_retries == 5
-
-    def test_roundtrip_serialization(self, tools, monkeypatch):
-        """Test that serialization and deserialization produces equivalent component."""
-        monkeypatch.setenv("MISTRAL_API_KEY", "test-key")
-
-        original = MistralAgentSDK(
-            agent_id="ag-roundtrip",
-            tools=tools,
-            tool_choice="any",
-            generation_kwargs={"max_tokens": 300, "random_seed": 123},
-        )
-
-        data = original.to_dict()
-        restored = MistralAgentSDK.from_dict(data)
-
-        assert restored.agent_id == original.agent_id
-        assert restored.tool_choice == original.tool_choice
-        assert restored.generation_kwargs == original.generation_kwargs
-        assert len(restored.tools) == len(original.tools)
-
-
-# =============================================================================
-# PIPELINE INTEGRATION TESTS
-# =============================================================================
-
-
-class TestMistralAgentSDKPipeline:
-    """Tests for pipeline integration."""
-
-    def test_pipeline_serialization(self, monkeypatch):
-        """Test that the component can be serialized in a pipeline."""
-        monkeypatch.setenv("MISTRAL_API_KEY", "test-key")
-
-        pipeline = Pipeline()
-        pipeline.add_component("agent", MistralAgentSDK(agent_id="ag-pipeline"))
-
-        # Serialize to dict
-        pipeline_dict = pipeline.to_dict()
-
-        assert "agent" in pipeline_dict["components"]
-        assert (
-            pipeline_dict["components"]["agent"]["type"]
-            == "haystack_integrations.components.agents.mistral.agent_sdk.MistralAgentSDK"
-        )
-
-        # Deserialize
-        restored_pipeline = Pipeline.from_dict(pipeline_dict)
-        restored_agent = restored_pipeline.get_component("agent")
-
-        assert restored_agent.agent_id == "ag-pipeline"
-
-    def test_pipeline_yaml_roundtrip(self, monkeypatch):
-        """Test YAML serialization and deserialization."""
-        monkeypatch.setenv("MISTRAL_API_KEY", "test-key")
-
-        pipeline = Pipeline()
-        pipeline.add_component(
-            "agent",
-            MistralAgentSDK(
-                agent_id="ag-yaml",
-                generation_kwargs={"max_tokens": 100}
-            )
-        )
-
-        yaml_str = pipeline.dumps()
-        restored = Pipeline.loads(yaml_str)
-
-        agent = restored.get_component("agent")
-        assert agent.agent_id == "ag-yaml"
-        assert agent.generation_kwargs == {"max_tokens": 100}
-
-
-# =============================================================================
-# ERROR HANDLING TESTS
-# =============================================================================
-
-
-class TestMistralAgentSDKErrors:
-    """Tests for error handling."""
-
-    @patch("haystack_integrations.components.agents.mistral.agent_sdk.MistralAgentSDK.warm_up")
-    def test_sdk_import_error(self, mock_warm_up, monkeypatch):
-        """Test handling of missing mistralai package."""
-        monkeypatch.setenv("MISTRAL_API_KEY", "test-key")
-
-        agent = MistralAgentSDK(agent_id="ag-test")
-        agent._client = None  # Force re-initialization
-
-        with patch.dict("sys.modules", {"mistralai": None}):
-            # This would raise ImportError when warm_up tries to import
-            # We test the real warm_up behavior
-            pass  # The actual import test would need a real missing module
-
-    @patch("haystack_integrations.components.agents.mistral.agent_sdk.MistralAgentSDK.warm_up")
-    def test_api_error_handling(self, mock_warm_up, chat_messages, monkeypatch):
-        """Test that API errors are properly wrapped."""
-        monkeypatch.setenv("MISTRAL_API_KEY", "test-key")
-
-        agent = MistralAgentSDK(agent_id="ag-test")
-        agent._client = MagicMock()
-
-        # Create a mock error
-        class MockMistralError(Exception):
-            def __init__(self):
-                self.status_code = 422
-                self.message = "Invalid agent_id"
-
-        agent._client.agents.complete.side_effect = MockMistralError()
-
-        # The error should be raised (exact handling depends on implementation)
-        with pytest.raises(Exception):
-            agent.run(chat_messages)
-
